search frontend: move monacoToRange
