+++ client/shared/src/search/query/completion.ts
@@ -1,7 +1,8 @@
 import * as Monaco from 'monaco-editor'
 import { escapeRegExp, startCase } from 'lodash'
 import { FILTERS, resolveFilter } from './filters'
-import { Token, toMonacoRange } from './scanner'
+import { Token } from './scanner'
+import { toMonacoRange } from './tokens'
 import { Omit } from 'utility-types'
 import { Observable } from 'rxjs'
 import { IRepository, IFile, ISymbol, ILanguage, IRepoGroup } from '../../graphql/schema'
