+++ client/shared/src/search/query/tokens.ts
@@ -484,3 +484,13 @@ const toMonaco = (token: Token): Monaco.languages.IToken[] => {
  */
 export const getMonacoTokens = (tokens: Token[], toDecorate = false): Monaco.languages.IToken[] =>
     toDecorate ? tokens.flatMap(token => decorate(token).map(decoratedToMonaco)) : tokens.flatMap(toMonaco)
+
+/**
+ * Converts a zero-indexed, single-line {@link CharacterRange} to a Monaco {@link IRange}.
+ */
+export const toMonacoRange = ({ start, end }: CharacterRange): Monaco.IRange => ({
+    startLineNumber: 1,
+    endLineNumber: 1,
+    startColumn: start + 1,
+    endColumn: end + 1,
+})
